PP-2586 Renamed pay-apps to pay-products
With @kakumara
