+++ build-local.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-mvn -DskipTests clean package && docker build -t govukpay/payapps:local .
+mvn -DskipTests clean package && docker build -t govukpay/products:local .
