+++ src/main/java/uk/gov/pay/products/config/ProductsModule.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.apps.config;
+package uk.gov.pay.products.config;
 
 import com.codahale.metrics.MetricRegistry;
 import com.google.inject.AbstractModule;
@@ -7,10 +7,10 @@ import io.dropwizard.setup.Environment;
 
 public class ProductsModule extends AbstractModule {
 
-    final PayAppsConfiguration configuration;
+    final ProductsConfiguration configuration;
     final Environment environment;
 
-    public ProductsModule(PayAppsConfiguration configuration, Environment environment) {
+    public ProductsModule(ProductsConfiguration configuration, Environment environment) {
         this.configuration = configuration;
         this.environment = environment;
     }
